Move configuration to separate file for easier reuse
